requirements.txt
hydra_colorlog==1.2.0
hydra-core==1.3.2
jams>=0.3.4
jupyterlab
librosa==0.10.0
matplotlib
mido>=1.2.10
mir_eval>=0.7
mirdata>=0.3.6
omegaconf==2.3.0
pandas>=1.4.2
pydub==0.25.1
requests>=2.27.1
scipy>=1.8.1
sounddevice>=0.4.4
tensorboard>=2.12.0
tensorboardX>=2.5
torch==2.0.0 
torchaudio==2.0.1
torchinfo
torchsummary==1.5.1
tqdm==4.65.0
sed_eval
stickytape
git+git://github.com/lehduong/torch-warmup-lr.git